+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Updated tools:
 Newly installed tools:
 
  * installs VirtualBox 5.1.18 and make it the default vagrant provider (see [PR #28](https://github.com/tknerr/linus-kitchen/pull/28))
+ * installs Packer 1.0.0 (see [PR #36](https://github.com/tknerr/linus-kitchen/pull/36))
 
 
 ## 0.1 (May 11, 2016)